feat: implement organization branding management with upload and link options
- Add Branding component for managing organization logos.
- Create UploadTabContent and LinkTabContent for respective branding methods.
- Implement BrandingPreview for visualizing logo changes.
- Add unsaved changes dialog to prevent accidental loss of changes.
- Introduce useBranding hook for handling logo save and remove operations.
- Enhance utility functions with isValidHttpUrl for URL validation.
- Write tests for Branding component to ensure functionality and user interactions.
