+++ app/settings/branding/page.tsx
@@ -1,4 +1,4 @@
-import Branding from '@/components/settings/branding';
+import Branding from '@/components/settings/branding/Branding';
 
 export default function SettingsBrandingPage() {
   return <Branding />;
